Hacky fix for ships freezing after firing bullet.
An issue has been present in the emulator where when you fire a
bullet, the ship freezes. However, all the game state is still
because if you move and fire, the bullet comes from the location
where you should be. This issue is only present in the emulator,
not when actually booting into the game.

The fix involved moving the ship up 8 rows on the screen to avoid
being in the last 8 rows, where nothing was getting redrawn.