Partial bug fixes for write behind
Write behind and read ahead break a lot of concurrency things.
There have been a few bugs that have been ironed out but they
are still very broken. In particular, they acquire a lock on the
sector, go to write the sector to disk, and never come back from
there. This is an issue because it never releases the lock.